refactor(discord): move runtime session gc helpers
